Refactored into two folders
