+++ frontend/src/scenes/session-recordings/player/inspector/playerInspectorLogic.test.ts
@@ -1,18 +1,18 @@
 import { initKeaTests } from '~/test/init'
 import { expectLogic } from 'kea-test-utils'
-import { sharedListLogic } from 'scenes/session-recordings/player/inspector/sharedListLogic'
+import { playerInspectorLogic } from 'scenes/session-recordings/player/inspector/playerInspectorLogic'
 import { RecordingWindowFilter } from '~/types'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 
 const playerLogicProps = { sessionRecordingId: '1', playerKey: 'playlist' }
 
-describe('sharedListLogic', () => {
-    let logic: ReturnType<typeof sharedListLogic.build>
+describe('playerInspectorLogic', () => {
+    let logic: ReturnType<typeof playerInspectorLogic.build>
 
     beforeEach(() => {
         initKeaTests()
         featureFlagLogic.mount()
-        logic = sharedListLogic(playerLogicProps)
+        logic = playerInspectorLogic(playerLogicProps)
         logic.mount()
     })
 
@@ -37,28 +37,6 @@ describe('sharedListLogic', () => {
                 .toMatchValues({
                     windowIdFilter: RecordingWindowFilter.All,
                 })
-        })
-    })
-
-    describe('miniFilters', () => {
-        it('should start with the first entry selected', async () => {
-            expect(logic.values.selectedMiniFilters).toEqual([
-                'all-automatic',
-                'console-all',
-                'events-all',
-                'performance-all',
-            ])
-        })
-
-        it('should remove other selected filters if alone', async () => {
-            logic.actions.setMiniFilter('all-errors', true)
-
-            expect(logic.values.selectedMiniFilters.sort()).toEqual([
-                'all-errors',
-                'console-all',
-                'events-all',
-                'performance-all',
-            ])
         })
     })
 })
