+++ frontend/src/scenes/session-recordings/player/inspector/v1/ConsoleDetails.tsx
