+++ frontend/src/scenes/session-recordings/player/inspector/v2/components/ItemConsoleLog.tsx
@@ -1,8 +1,8 @@
 import { LemonButton } from '@posthog/lemon-ui'
-import { SharedListItemConsole } from '../../sharedListLogic'
+import { InspectorListItemConsole } from '../../playerInspectorLogic'
 
 export interface ItemConsoleLogProps {
-    item: SharedListItemConsole
+    item: InspectorListItemConsole
     expanded: boolean
     setExpanded: (expanded: boolean) => void
 }
@@ -11,12 +11,12 @@ export function ItemConsoleLog({ item, expanded, setExpanded }: ItemConsoleLogPr
     return (
         <>
             <LemonButton noPadding onClick={() => setExpanded(!expanded)} status={'primary-alt'} fullWidth>
-                <div className="p-2 text-xs cursor-pointer truncate font-mono">{item.data.previewContent}</div>
+                <div className="p-2 text-xs cursor-pointer truncate font-mono">{item.data.content}</div>
             </LemonButton>
 
             {expanded && (
                 <div className="p-2 text-xs border-t">
-                    <pre className="whitespace-pre-wrap">{item.data.fullContent}</pre>
+                    <pre className="whitespace-pre-wrap">{item.data.content}</pre>
                 </div>
             )}
         </>
