+++ frontend/src/scenes/session-recordings/player/playerSettingsLogic.test.ts
@@ -1,12 +1,16 @@
 import { expectLogic } from 'kea-test-utils'
+import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { initKeaTests } from '~/test/init'
 import { playerSettingsLogic } from './playerSettingsLogic'
 
 describe('playerSettingsLogic', () => {
     let logic: ReturnType<typeof playerSettingsLogic.build>
+    let ffLogic: ReturnType<typeof featureFlagLogic.build>
 
     beforeEach(() => {
         initKeaTests()
+        ffLogic = featureFlagLogic()
+        ffLogic.mount()
         logic = playerSettingsLogic()
         logic.mount()
     })
@@ -46,6 +50,73 @@ describe('playerSettingsLogic', () => {
                 .toMatchValues({
                     showOnlyMatching: true,
                 })
+        })
+    })
+
+    describe('miniFilters', () => {
+        afterEach(() => {
+            localStorage.clear()
+        })
+        it('should start with the first entry selected', async () => {
+            expect(logic.values.selectedMiniFilters).toEqual([
+                'all-automatic',
+                'console-all',
+                'events-all',
+                'performance-all',
+            ])
+        })
+
+        it('should remove other selected filters if alone', async () => {
+            logic.actions.setMiniFilter('all-errors', true)
+
+            expect(logic.values.selectedMiniFilters.sort()).toEqual([
+                'all-errors',
+                'console-all',
+                'events-all',
+                'performance-all',
+            ])
+        })
+
+        it('should allow multiple filters if not alone', async () => {
+            logic.actions.setMiniFilter('console-warn', true)
+            logic.actions.setMiniFilter('console-info', true)
+
+            expect(logic.values.selectedMiniFilters.sort()).toEqual([
+                'all-automatic',
+                'console-info',
+                'console-warn',
+                'events-all',
+                'performance-all',
+            ])
+        })
+
+        it('should reset to first in tab if empty', async () => {
+            expect(logic.values.selectedMiniFilters.sort()).toEqual([
+                'all-automatic',
+                'console-all',
+                'events-all',
+                'performance-all',
+            ])
+            logic.actions.setMiniFilter('console-warn', true)
+            logic.actions.setMiniFilter('console-info', true)
+
+            expect(logic.values.selectedMiniFilters.sort()).toEqual([
+                'all-automatic',
+                'console-info',
+                'console-warn',
+                'events-all',
+                'performance-all',
+            ])
+
+            logic.actions.setMiniFilter('console-warn', false)
+            logic.actions.setMiniFilter('console-info', false)
+
+            expect(logic.values.selectedMiniFilters.sort()).toEqual([
+                'all-automatic',
+                'console-all',
+                'events-all',
+                'performance-all',
+            ])
         })
     })
 })
